Split the scheduler into multiple mixin classes to reduce the file size (#8483)
